refactor: move tests package
