+++ src/test/kotlin/lotto/domain/LottoResultTest.kt
@@ -1,8 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
-import lotto.domain.LottoResult
-import lotto.domain.WinningLotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
