+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
