+++ src/test/kotlin/lotto/domain/LottosTest.kt
@@ -1,7 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
-import lotto.domain.Lottos
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
