+++ src/test/kotlin/lotto/domain/WinningLottoTest.kt
@@ -1,7 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
-import lotto.domain.WinningLotto
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
